requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain-community
openai
tiktoken
langchain-openai
python-dotenv
langchain-chroma